Update from_rational to support long long parameters
Changed the parameter types of from_rational function from long to long long in type.h and value.h for better handling of larger rational numbers. This ensures compatibility and consistency across different parts of the code.

+++ platform/include/roughpy/devices/type.h
@@ -139,7 +139,7 @@ struct NumberTraits {
 
     PowFn pow = nullptr;
 
-    void (*from_rational)(void*, long, long) = nullptr;
+    void (*from_rational)(void*, long long, long long) = nullptr;
 };
 
 using HashFn = hash_t (*)(const void*);
